requirement.txt
APScheduler==3.5.3
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
cx-Oracle==7.0.0
decorator==4.3.0
dominate==2.3.5
Flask==1.0.2
Flask-APScheduler==1.10.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
pbr==5.1.1
pycparser==2.19
PyMySQL==0.9.2
python-dateutil==2.7.5
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
spyne==2.12.14
SQLAlchemy==1.2.12
sqlalchemy-migrate==0.11.0
sqlparse==0.2.4
suds-jurko==0.6
Tempita==0.5.2
tzlocal==1.5.1
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
